Issue #265 Refactor import counter to count elements for more than one input file
