+++ bc211/importer.py
@@ -12,43 +12,9 @@ from bc211.exceptions import XmlParseException
 LOGGER = logging.getLogger(__name__)
 
 
-class ImportCounters:
-    def __init__(self):
-        self.organization_count = 0
-        self.location_count = 0
-        self.service_count = 0
-        self.taxonomy_term_count = 0
-        self.address_count = 0
-        self.phone_number_types_count = 0
-        self.phone_at_location_count = 0
-
-    def count_organization(self):
-        self.organization_count += 1
-
-    def count_location(self):
-        self.location_count += 1
-
-    def count_service(self):
-        self.service_count += 1
-
-    def count_taxonomy_term(self):
-        self.taxonomy_term_count += 1
-
-    def count_address(self):
-        self.address_count += 1
-
-    def count_phone_number_types(self):
-        self.phone_number_types_count += 1
-
-    def count_phone_at_location(self):
-        self.phone_at_location_count += 1
-
-
-def save_records_to_database(organizations):
+def save_records_to_database(organizations, counters):
     translation.activate('en')
-    counters = ImportCounters()
     save_organizations(organizations, counters)
-    return counters
 
 
 def save_organizations(organizations, counters):
