+++ bc211/management/commands/import_bc211_data.py
@@ -2,9 +2,11 @@ import argparse
 from django.core.management.base import BaseCommand
 from bc211.parser import read_records_from_file
 from bc211.importer import save_records_to_database
+from bc211.import_counters import ImportCounters
 
 # invoke as follows:
 # python manage.py import_bc211_data path/to/bc211.xml
+
 
 class Command(BaseCommand):
     help = 'Import BC-211 data from XML file'
@@ -22,7 +24,8 @@ class Command(BaseCommand):
         records = read_records_from_file(file)
 
         self.stdout.write(self.style.SUCCESS('Writing data to database'))
-        counts = save_records_to_database(records)
+        counts = ImportCounters()
+        save_records_to_database(records, counts)
 
         message_template = ('Successfully imported {0} organization(s), '
                             '{1} location(s), {2} service(s), '
